[비동기 댓글 기능] 티버 미션 제출합니다. (#203)
* Revert "revert to home"

This reverts commit 1ee97e3d7aff3d1161d08f942f375d4f15dd0ac8.

* [게시글 생성/조회/수정/삭제] 버디 미션 제출합니다. (#26)

* create README

* feat : 게시글 작성 페이지 이동

* feat: 게시글 작성 기능

게시글 작성 기능 구현
contents 마크다운으로 보여주도록

* feat: 게시글 목록 조회 기능

* feat: 게시글 조회 기능추가

* refactor: 리팩토링 및 테스트 구현

* style: 불필요한 코드 제거

* refactor: submit_article redirect로 변경 및 테스트코드 수정

* feat: Article 수정 메서드 및 테스트 구현

* refactor: Article,ArticleRepository 리팩토링

* feat:게시글 삭제 구현 및 테스트 추가

* refactor: view의 중복코드 제거(thymeleaf)

* refactor: view의 article page 중복코드 제거(thymeleaf)

* style:안쓰는 코드 제거

* refactor: 마크다운 문법 수정 적용으로 수정

* style: 컨벤션 수정

* fix:DI 생성자 수정

* refactor:ArticleControllerTest 중복코드 리팩토링

* refactor:ArticleRepositoryTest 리팩토링

* refactor: 피드백 수정

* feat: ArticleTest 추가

* fix: Test코드 추상화 및 Article class의 incresement변수 동기화문제 리팩토링

* refactor: 필요없는 sout문 삭제

* fix: ArticleControllerTest의 필드주입을 생성자 주입으로 변경

* [게시판 회원 기능] 버디 미션 제출 합니다. (#67)

* create README

* feat : 게시글 작성 페이지 이동

* feat: 게시글 작성 기능

게시글 작성 기능 구현
contents 마크다운으로 보여주도록

* feat: 게시글 목록 조회 기능

* feat: 게시글 조회 기능추가

* refactor: 리팩토링 및 테스트 구현

* style: 불필요한 코드 제거

* refactor: submit_article redirect로 변경 및 테스트코드 수정

* feat: Article 수정 메서드 및 테스트 구현

* refactor: Article,ArticleRepository 리팩토링

* feat:게시글 삭제 구현 및 테스트 추가

* refactor: view의 중복코드 제거(thymeleaf)

* refactor: view의 article page 중복코드 제거(thymeleaf)

* style:안쓰는 코드 제거

* refactor: 마크다운 문법 수정 적용으로 수정

* style: 컨벤션 수정

* fix:DI 생성자 수정

* refactor:ArticleControllerTest 중복코드 리팩토링

* refactor:ArticleRepositoryTest 리팩토링

* refactor: 피드백 수정

* feat: ArticleTest 추가

* fix: Test코드 추상화 및 Article class의 incresement변수 동기화문제 리팩토링

* refactor: 필요없는 sout문 삭제

* fix: ArticleControllerTest의 필드주입을 생성자 주입으로 변경

* fix: JPA로 인한 코드 수정

* refactor: Controller uri 중복제거

* docs: README.md 2주차 기능 명세

* feat: mySql 데이터베이스로 변경

* feat: 회원등록 기능 구현 및 테스트 메서드 추가

* feat: 회원조회 기능 구현 및 테스트 추가

* feat: 회원등록 이메일 중복확인 구현 및 테스트 추가

* feat: html 레이아웃 적용

* feat: signup 중복이메일 메시지 출력 구현

* feat: UserDto 추가
User 유효성 검사 (Front, Back) 및 테스트 추가

* feat: 프론트에서 비밀번호 확인 제출 불가능하게 구현

* feat: lombok 적용, 프론트 uri 수정

* feat: 로그인할 때 이메일 비밀번호 확인

* refactor: nav-bar 중복제거

* feat: logout 기능 구현 및 테스트 추가

* feat: 마이페이지로 이동 구현 및 테스트 추가

* feat: Validator 추가

* feat: 회원 수정 기능 구현

* feat: 회원 삭제 기능 구현

* feat: 예외 처리 추가

* feat: ExceptionHandler 구현

* feat: ExceptionHandler 테스트 구현(UserControllerTest만)

* feat: 로그인테스트 구현 및 예외 추가

* refactor: UserController 의 비즈니스로직 UserService로 분리

* refactor: LoginController 의 비즈니스로직 LoginService로 분리

* refactor: UserService의 session로직 컨트롤러로 이동

* style: 패키지 구조 정리

* fix: 전체 로직 확인 및 수정

* feat: Article의 toEntity 메서드 추가 및 ArticleController의 비즈니스 로직 Service계층으로 분리

* refactor: LoginControllerTest,UserControllerTest 중복코드 메서드 분리

* refactor: UserControllerTest 중복코드 메서드 분리

* refactor: 전체 코드 리팩토링

* fix: 불필요한 코드 삭제

* feat: UserDto 테스트 추가

* feat:sessionTest 추가

* feat:userServiceTest 추가

* fix:UserControllerTest sessionTest 수정

* feat: ArticleServiceTest 코드 추가

* [댓글 기능] 버디 미션 제출합니다.  (#118)

* docs: 댓글 기능 명세

* feat: Interceptor구현

* fix: Controller,Service HttpSession 중복 제거

* fix: Interceptor 기능 수정

* feat: Article과 User의 관계 매핑 및 Interceptor 적용

* feat: comment 작성 구현

* feat: Comment 삭제 기능 구현

* feat: Comment 수정 기능 구현

* feat: ArticleControllerTest 테스트 케이스 추가 및 data.sql 추가

* feat: data.sql 추가

* feat: ArticleTest 추가

* refactor: ArticleController의 댓글 로직 ConmmentController로 분리 및 테스트 이동

* feat: comment 기능 추가 및 테스트 추가

* feat: Article,Comment Class의 contents필드에 @Lob어노테이션 추가

* refactor: ArticleController, ArticleService 리팩토링 및 테스트 추가

* feat: Date Class 생성 및 객체 상속

* refactor: 테스트에 사용되는 공용 필드 static으로 변경 및 리팩토링

* fix: LoginService의 login 메서드 filter로 변경 및 테스트 코드 추가

* refactor: 테스트 코드 전체 리팩토링

* feat: DataJpaTest 추가

* refactor: ExceptionHandler 메서드 공통 핸들러 클래스로 분리

* refactor: 피드백반영

* docs: README 기능 명세

* feat: 댓글 생성기능 ajax 구현

* feat: 댓글 수정기능 ajax 구현

* feat: 댓글 삭제기능 ajax 구현

* test: ajax 댓글 기능 테스트 구현

* style: 컨밴션 확인 및 불필요한 주석 제거

* refactor: sql 파일 수정

* refactor: properties 파일 수정

* refactor: article.html 버그 수정

* feat: 스크립트 작성

* refactor: 패키지 구조 수정

* refactor: sql 수정

* feat: logback 추가

* refactor: request url 수정

* fix: 마지막 댓글 삭제 시 오류 수정

* refactor: checkAvailableUserForDelete 메소드 삭제

* refactor: Comment Controller 반환 값 CommentDto로 수정

* refactor: ArticleController에서 반환 값 ArticleDto 반환하도록 수정

* feat: 현재 세션 값 확인하는 UserSession 어노테이션 구현

-LoginUser 생성

* refactor: UserDto 생성자 추가

* refactor: checkCorrespondingAuthor 메소드의 매개변수 수정

* refactor: matchEmail 메소드 생성

* refactor: LoginUser 사용하도록 수정

* refactor: 코드 컨벤션을 위한 리팩토링

* refactor: CommentResponseDto 생성

* refactor: ArticleResponseDto 생성

* refactor: 코드 컨벤션을 위한 리팩토링

+++ src/main/java/techcourse/myblog/service/CommentService.java
@@ -1,65 +1,45 @@
 package techcourse.myblog.service;
 
-import techcourse.myblog.domain.article.Article;
-import techcourse.myblog.domain.comment.Comment;
-import techcourse.myblog.domain.comment.Contents;
-import techcourse.myblog.domain.user.User;
-import techcourse.myblog.repository.CommentRepository;
-import techcourse.myblog.service.exception.NotFoundCommentException;
-import techcourse.myblog.service.exception.UnauthorizedException;
-
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
+import techcourse.myblog.domain.comment.Comment;
+import techcourse.myblog.domain.comment.CommentRepository;
+import techcourse.myblog.domain.dto.response.LoginUser;
+import techcourse.myblog.exception.NotFoundObjectException;
+import techcourse.myblog.domain.dto.CommentDto;
+import techcourse.myblog.domain.user.User;
+
+import javax.transaction.Transactional;
 
 @Service
 public class CommentService {
-	private CommentRepository commentRepository;
 	private UserService userService;
-	private ArticleService articleService;
+	private CommentRepository commentRepository;
 
-	public CommentService(CommentRepository commentRepository, UserService userService, ArticleService articleService) {
-		this.commentRepository = commentRepository;
+	public CommentService(UserService userService, CommentRepository commentRepository) {
 		this.userService = userService;
-		this.articleService = articleService;
+		this.commentRepository = commentRepository;
 	}
 
-	public Comment save(User loginUser, Long articleId, Contents contents) {
-		User user = userService.findUser(loginUser);
-		Article article = articleService.findById(articleId);
-		Comment comment = new Comment(user, contents);
-		article.addComment(comment);
+	public Comment createComment(CommentDto commentDto, LoginUser loginUser) {
+		User user = userService.findByEmail(loginUser.getEmail());
+		Comment comment = commentDto.toEntity(user);
 		return commentRepository.save(comment);
 	}
 
-	@Transactional
-	public void update(User loginUser, Long articleId, Long commentId, Contents contents) {
-		Comment comment = findById(commentId);
-		existArticle(articleId);
-		confirmAuthorization(loginUser, comment.getAuthor());
-		comment.update(contents);
-	}
-
-	public Comment findById(Long commentId) {
-		return commentRepository.findById(commentId)
-				.orElseThrow(NotFoundCommentException::new);
-	}
-
-	private void existArticle(Long articleId) {
-		articleService.findById(articleId);
-	}
-
-	private void confirmAuthorization(User loginUser, User commentAuthor) {
-		User user = userService.findUser(loginUser);
-		if (!user.matchUser(commentAuthor)) {
-			throw new UnauthorizedException();
-		}
-	}
-
-	@Transactional
-	public void delete(User loginUser, Long articleId, Long commentId) {
-		Comment comment = findById(commentId);
-		confirmAuthorization(loginUser, comment.getAuthor());
-		existArticle(articleId);
+	public void deleteComment(Long commentId, LoginUser loginUser) {
+		User user = userService.findByEmail(loginUser.getEmail());
+		Comment comment = commentRepository.findById(commentId).orElseThrow(NotFoundObjectException::new);
+		comment.checkCorrespondingAuthor(user);
 		commentRepository.deleteById(commentId);
+	}
+
+	@Transactional
+	public Comment updateComment(Long commentId, LoginUser loginUser, CommentDto commentDto) {
+		User user = userService.findByEmail(loginUser.getEmail());
+		Comment comment = commentRepository.findById(commentId).orElseThrow(NotFoundObjectException::new);
+		Comment updateComment = commentDto.toEntity(user);
+		comment.update(updateComment, user);
+		return comment;
 	}
 }
